Implementação e ajustes consultas eletrodomesticos

+++ src/main/java/com/code4/voltz/repositorio/EnderecoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
-    List<Endereco> findByUsuarioId(long id);
+    List<Endereco> findByUsuarioId(int id);
     List<Endereco> findByRuaAndNumero(String rua, String numero);
 }
